libstagefright: Make it possible to skip OMX buffer reallocation
Some devices don't like the call to setParameter() at this point, so
skip this call if enough buffers are already allocated. This check
was present in KitKat but got removed when code to allocate extra-
buffers was introduced.

Change-Id: I33fd11fc269a649bb20bbdfd205dbfbb9a9902ff
Soong requires different handling, so instead of enabling this based on
TARGET_BOARD_PLATFORM, use dedicated BOARD_CANT_REALLOCATE_OMX_BUFFERS.
Signed-off-by: D. Andrei Măceș <andrei@unlegacy-android.org>